Armor
I added armor and code as well.

+++ DrakeShot-main/Drake Shot/Assets/scripts/Player/BulletScript.cs
@@ -15,7 +15,7 @@ public class Bullet : MonoBehaviour {
         
     }
     
-    public void Setup(float speed, int damage) {
+    public void Setup(int speed, int damage) {
 
         _rb.linearVelocity = transform.right * speed;
